+++ odoo14_sumex_apps_prestashop_export_products/__manifest__.py
@@ -1,0 +1,20 @@
+{
+    'name': 'Sumex Prestashop export products',
+    'description': 'Sumex Prestashop export products',
+    'author': 'Sumex S.A.',
+    'license': "Other proprietary",
+    "version": "15.0",
+    'depends': [
+        'base',
+        # 'contacts',
+        # 'account',
+        'stock',
+        # 'sale_management',
+        # 'purchase'
+    ],
+    'data': [
+        'views/forms_inherit_product_template.xml',
+    ],
+    'application': False,
+    'installable': True,
+}