+++ odoo_multi_channel_sale/changelog.md
@@ -6,10 +6,43 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [2.3.12] - 2021-10-29
+
+### Fixed
+- Product Write
+
+
+## [2.3.11] - 2021-05-28
+
+### Fixed
+- Default placeholder image render
+## Added
+- New alias image name controller
+
+
+## [2.3.9] - 2021-04-07
+
+### Fixed
+- Instance Kanban View
+
+
+## [2.3.8] - 2021-04-06
+
+### Fixed
+- Get Tax Ids from tax lines in order feed
+
+
+## [2.3.7] - 2021-04-05
+
+### Fixed
+- Window Actions
+
+
 ## [2.3.6] - 2021-01-23
 
 ### Fixed
 - Product/Partner/Order form view
+
 
 ## [2.3.5] - 2021-01-18